Additions/solaris/SharedFolders: Fix for incorrect stat association with dirents while reading large directories.
Fix nasty bug when reading large directories. The dirent's and stats were
located in two separate, parallel buffers; but when calling readdir with
a non-zero offset we were skipping through only the dirent's, thereby ending
up associating the wrong stats with these dirents.

This fix merges the two buffers into one structure, thereby both eliminating
the possibility of this bug and also simplifying the code.

Also included is proper validation of the offset that is passed in to
readdir.

Plus other minor fixes.


svn:sync-xref-src-repo-rev: r74790