bugref:8748: Additions/Graphics/Wayland: investigate EGLStreams support feasibility
Fix glXChooseFBConfig not to bail out if GLX_BUFFER_SIZE is requested, as this is ignored according to the specification for RGBA rendering, which is the only type that we support.  This lets glmark2 run for a bit until it exits due to a missing glGenerateMipmap symbol which is only core as of OpenGL 3.0.  We get a number of assersions in CrFbEntryRegionsSet() on the host in-between: Assert(!hEntry->Flags.fInList).

svn:sync-xref-src-repo-rev: r113224