bugref:8524: Additions/linux: play nicely with distribution-installed Additions Remove old VbglR3InterruptEventWaits() leftover from seamless.cpp.
Since the VbglR3InterruptEventWaits() API in VBoxGuest is rather clumsy, we
decided to forbid use of VbglR3WaitEvent() in a VBoxGuest session after
VbglR3InterruptEventWaits() has been called.  This change removes the last
use of it (there were only two in the whole tree) which conflicts with that:
a sanity test in seamless.cpp which is testing something no longer relevant
in the current code.

svn:sync-xref-src-repo-rev: r116718