bugref:3809: Linux installer maintenance Clean up unneeded kernel modules when the vboxdrv service is started.
When a Linux kernel is removed on the host for which VirtualBox driver
modules were installed, the kernel directory and those modules were left
in place.  This change adds a check to the vboxdrv service script to look
for such orphaned modules on start-up and remove them along with the
directories.

svn:sync-xref-src-repo-rev: r117047