Additions/linux/installer: fix log rotation in Additions installer. bugref:3809: Linux installer maintenance
In a previous change set we changed log rotation to only create a new log
file on the first log statement.  Unfortunately, due to a wrongly
initialised variable our check for the first log statement always failed,
so that instead of rotating log files we continually appended to the same
one on successive runs.  Fix that.

svn:sync-xref-src-repo-rev: r118103