Additions/linux/shfl: stop supporting old host/mount.vboxsf versions. bugref:4567: Linux kernel driver maintenance
Based on [PATCH] Remove support for old hosts and userspace mount.vboxsf
Signed-off-by: Hans de Goede <hdegoede@redhat.com>

Versions of VirtualBox many years ago used a different host/guest protocol
for accessing shared folders, as well as a different guest kernel to
user-space mount helper protocol.  While we try to support old Additions
versions on the host side, we only aim to support old host versions in the
Additions to a limited extent, and the guest kernel interface is also only
fixed for a defined range of releases.

svn:sync-xref-src-repo-rev: r118458