Host drivers/Linux: use KERNEL_DS, not get_ds() in VBoxPci. bugref:4567: Linux kernel driver maintenance. Kernel commit 736706bee3298208343a76096370e4f6a5c55915 Author: Linus Torvalds <torvalds@linux-foundation.org> Date:   Mon Mar 4 10:39:05 2019 -0800
    get rid of legacy :get_ds(): function
    
    Every in-kernel use of this function defined it to KERNEL_DS (either as
    an actual define, or as an inline function).  It:s an entirely
    historical artifact, and long long long ago used to actually read the
    segment selector valueof :%ds: on x86.
    
    Which in the kernel is always KERNEL_DS.

No kernel release versioning on the patch, as the define was in the tree
when Linus did the initial import into git for v2.6.12-rc2 back in April 2005.
Note (Michael): checking shows that it was present back to 2.4.0 at least.
Thank you Valdis Klētnieks for the patch.

svn:sync-xref-src-repo-rev: r129366