IPRT/VFS: 'VBoxManage import VM.ova' can fail due to reading some of the data twice. bugref:9730
Attempting to import a VM may fail on hosts such as Solaris which use
the Runtime/posix/r3 implementation of RTFileReadAt() or RTFileWriteAt().
These routines on POSIX platforms leverage pread(2) and pwrite(2)
respectively which don't update the file-position indicator which means
callers need to take this into account.  The appliance import code could
run into this when rtVfsStdFile_Read() is called with a non-negative
offset.


svn:sync-xref-src-repo-rev: r147896