VBoxGuest-netbsd.c: Fix mouse range. bugref:8547
wsmux(4) doesn't properly handle the case when two mice with absolute
position events but different calibration data are being multiplexed.
Force VMM range into the ums(4) range and rely on the fact that no
actual calibration is done and both devices are used in the raw mode.


svn:sync-xref-src-repo-rev: r151563