Devices/EFI/Firmware: Fix broken TPM device with Linux guests after UEFI firmware update, github:gh-143
Presumably https://github.com/tianocore/edk2/commit/ce4c76e46d52e24551f4986bded4c9b764502200 broke the TPM device 
with Linux guests as the TPM driver wouldn't be able to map the MMIO region as ioremap() expects an uncachable memory region
while it got a writeback one instead, causing in an error (-12 -> ENOMEM) trying to initialize the driver.
The commit changed the default MTRR type from uncacheable to writeback which would cause the ACPI driver in Linux to reserve a
writeback memory region from 0xfed40000 - 0xfed41fff when parsing the OperationRegion() statement from vbox-tpm.dsl.
The fix is to explicitely mark the TPM MMIO region as uncachable when the TPM is present so the region is already populated in the MTRR 
registers.


svn:sync-xref-src-repo-rev: r170277