VMM/GCM,Main,Devices/PC/DevACPI,include/VBox: Introduce facility for patching a guest in order to work around certain issues, bugref:4017
This makes it possible to add guest specific patches and trigger the patching process from devices to workaround guest issues.
The first issue being worked around is the infamous Linux panic happening on older guests if an IO-APIC is configured because the
function validating that timer interrupts work has too tight limits for a virtualized environment. This should reduce test failure
rate on certain hosts. The patching is disabled by default and needs to be enabled on a per VM basis.



svn:sync-xref-src-repo-rev: r170644